Fix sensor setup for battery (dis)charging power
* `CurBatChaArr.1` is a subcomponent of the overall value in
`CurBatCha`, thus it should have the `_a` identifier.
* Swap the `_a` and `_total_` identifiers for `CurBatDsch` and
`CurBatDschArr.1`, reflecting the subcomponent hierarchy as well.

In both cases, the `Arr.1` value is slightly different from the total
value for me, but it is not obvious why and how. I just have one
battery, so in my world they should be the same. Therefore I'd stick
to the hierarchical setup. If needed both can be made available in Home
Assistant.